Grand value division: value::index
